#15 extract change tracking logic to mongoose plugin
