+++ models/developer.js
@@ -2,18 +2,10 @@
 
 const request = require('request');
 const mongoose = require('mongoose');
-const _ = require('lodash');
-const moment = require('moment');
+const DateChangeTrackingPlugin = require('../plugins/date-change-tracking');
 const Schema = mongoose.Schema;
 
-const AvailableDateSchema = new Schema({
-    version: Number,
-    value: Date,
-    created_at: Date
-});
-
 const DeveloperSchema = new Schema({
-    availableDateHistory: [AvailableDateSchema],
     name: String,
     firstName: String,
     lastName: String,
@@ -32,32 +24,9 @@ const DeveloperSchema = new Schema({
     }
 });
 
-DeveloperSchema
-    .virtual('availableDate')
-    .get(function() {
-        let lastUpdated = _.sortBy(this.availableDateHistory, 'version').pop();
-        if(!lastUpdated) {
-            return null;
-        }
-        return lastUpdated.value;
-    })
-    .set(function(val) {
-        let lastUpdated = _.sortBy(this.availableDateHistory, 'version').pop() || {
-            version: 0,
-            value: null
-        };
-        let valueRemainedEmpty = val === null && lastUpdated.value === null;
-        let valueDidntChange = moment(val).isSame(lastUpdated.value, 'day');
-        if(valueRemainedEmpty || valueDidntChange) {
-            return;
-        }
-        let newVersionNumber = lastUpdated.version + 1;
-        this.availableDateHistory.push({
-            version: newVersionNumber,
-            created_at: new Date(),
-            value: val
-        });
-    });
+DeveloperSchema.plugin(DateChangeTrackingPlugin, {
+    field: 'availableDate'
+});
 
 Object.assign(DeveloperSchema.methods, {
     fetchLocation(address) {
